requirements.txt
streamlit
langchain
requests
beautifulsoup4
langchain-openai
langchain_openai
langchain_core
python-dotenv
langchain_community
langserve
fastapi
uvicorn
sse_starlette
beautifulsoup4
pypdf
chromadb
langchain_groq
langchain_ollama
faiss-cpu
groq
ollama
openai
langchain_openai
langchain-google-genai

